add if for destroy

+++ app/controllers/addresses_controller.rb
@@ -25,9 +25,12 @@ class AddressesController < ApplicationController
 
   def destroy
     @address = Address.find(params[:id])
-    if @address.destroy
-      redirect_back(fallback_location: root_path)
+    if @address
+      if @address.destroy
+        redirect_back(fallback_location: root_path)
+      end
     end
+    flash [:error] = "Une erreur est survenue"
   end
 
   private
