+++ app/controllers/avatars_controller.rb
@@ -7,13 +7,14 @@ class AvatarsController < ApplicationController
 
   def destroy
     @avatar = Cook.find(params[:cook_id]).avatar
-
-    if @avatar.destroy
-        redirect_to galleries_path, success: 'Avatar supprimé.'
-    else
-        render :show, error: 'Erreur, impossible de supprimer l\'avatar'
+    if @avatar
+      if @avatar.destroy
+          redirect_to galleries_path, success: 'Avatar supprimé.'
+      else
+          render :show, error: 'Erreur, impossible de supprimer l\'avatar'
+      end
     end
-    
+    flash [:error] = "Une erreur est survenue"
   end
 
 end
