+++ app/controllers/cook/photos_controller.rb
@@ -8,12 +8,14 @@ class Cook::PhotosController < ApplicationController
 
   def destroy
     @photo = Dish.find(params[:dish_id]).photo
-    
-    if @photo.destroy
-        redirect_to dishes_path, success: 'Photo supprimée.'
-    else
-        render :show, error: 'Erreur, impossible de supprimer la photo'
+    if @photo
+      if @photo.destroy
+          redirect_to dishes_path, success: 'Photo supprimée.'
+      else
+          render :show, error: 'Erreur, impossible de supprimer la photo'
+      end
     end
+    flash [:error] = "Une erreur est survenue"
   end
 
 end
