+++ app/controllers/order_dishes_controller.rb
@@ -39,16 +39,18 @@ class OrderDishesController < ApplicationController
   end
 
   def destroy 
-    if @order.destroy
-      flash[:success] = "Plat enlevé du panier avec succès."
-    else
-      flash[:error] = "Impossible de supprimer le plat."
-    end
-    respond_to do |format|
-      format.html{redirect_back(fallback_location: root_path)}
-      format.js{}
-    end
-
+    if @order
+      if @order.destroy
+        flash[:success] = "Plat enlevé du panier avec succès."
+      else
+        flash[:error] = "Impossible de supprimer le plat."
+      end
+      respond_to do |format|
+        format.html{redirect_back(fallback_location: root_path)}
+        format.js{}
+      end 
+    end  
+    flash [:error] = "Une erreur est survenue"
   end
   
   private
